T#798 P3-A: Extract settings module from server.ts
Phase 3 of T#764 (Library #102 modularization). Sub-task A — settings module,
2 handlers (GET + POST /api/settings), ~110 LoC inline in server.ts at 736-846.

Per the Phase 1 modularization pattern (T#769-T#784): extract handlers into
src/settings/routes.ts with a `registerSettingsRoutes` registrar, replace
inline handlers in server.ts with the registrar call, preserve behavior
exactly.

## Changes

**New: src/settings/routes.ts**
- Single registrar `registerSettingsRoutes(app, helpers)`
- SettingsHelpers interface: `getSetting`, `setSetting`, `logSecurityEvent`
  (passed via DI to keep the module testable)
- 2 handlers extracted verbatim:
  - GET /api/settings (no password hash exposed)
  - POST /api/settings (Gorn-only via UI; rejects beast API calls with
    impersonation_blocked security event; supports newPassword/removePassword/
    authEnabled/localBypass; logs settings_changed event)

**Modified: src/server.ts**
- Add `import { registerSettingsRoutes } from './settings/routes.ts'`
- Replace 110-line inline handler block (lines 736-846) with single
  `registerSettingsRoutes(app, { getSetting, setSetting, logSecurityEvent })`
- server.ts line count: 4428 → 4320 (-108 lines net)

## Order in server.ts

Registration call placed at the same approximate location as the prior inline
handlers (~line 737), well before the SPA fallback catch-all at line 4227.
Bug-class avoided: module-registration-must-precede-spa-catch-all (T#783 hotfix
lesson).

## Test plan

- [x] `bunx tsc --noEmit` clean on changed files. Zero new type errors
  introduced; pre-existing errors at src/server.ts:2531/2608/2614 (FormData
  upload handler) and 4081/4085 (OpenAPIHono AppEnv/Env mismatch) are
  unchanged and tracked separately.
- [x] Import-completeness: `registerSettingsRoutes` is the only new symbol
  reaching server.ts; the helpers (`getSetting`, `setSetting`,
  `logSecurityEvent`) are already imported in server.ts and just passed
  through.
- [ ] Runtime smoke at deploy gate:
  - `GET /api/settings` returns `{authEnabled, localBypass, hasPassword, vaultRepo}`
  - `POST /api/settings` with `?as=` returns 403 + `impersonation_blocked`
    security event logged
  - `POST /api/settings` from gorn-session updates settings + logs
    `settings_changed`

Reviewer: @zaghnal per task assignment.

## References

- T#798: sub-task of T#786 P3 (sibling under T#764 per Spec #56 v1 2-level
  constraint)
- T#786: parent phase task (Library #102 final cleanup)
- T#783: established the module-registration-must-precede-SPA-catch-all rule
- /denbook v1.0.8: bug-class registry referenced in T#798 description
